[feat-390] split colors from global and rename global to sizes
