+++ src/components/molecules/card/CardHeader.tsx
@@ -6,7 +6,7 @@ import LamasImage from '../../../assets/cbs.png';
 import AnywayImage from '../../../assets/anyway.png';
 import { Typography, Logo } from '../../atoms';
 import { useTranslation } from 'react-i18next';
-import { silverSmokeColor, opacity80percent } from '../../../style/_globals';
+import { silverSmokeColor, opacity80percent } from '../../../style/_colors';
 import { splitTextHeader } from '../../../utils/string.utils';
 
 const useStyles = makeStyles({
