+++ src/style/_sizes.ts
@@ -1,0 +1,22 @@
+export const defaultBorderRadius = 4;
+export const sidBarWidth = 300;
+// card
+export const cardHeight = 500;
+export const cardWidth = 500;
+// card content and actions (footer)
+export const cardHeaderHeight = 70;
+export const cardFooterHeight = 50;
+
+export const fontFamilyString = [
+  '"Alef"',
+  '-apple-system',
+  'BlinkMacSystemFont',
+  '"Segoe UI"',
+  'Roboto',
+  '"Helvetica Neue"',
+  'Arial',
+  'sans-serif',
+].join(',');
+
+//road number icon && font
+export const roadIconFont = 'sans-serif';
