+++ src/style/index.ts
@@ -1,2 +1,3 @@
-export * from './_globals';
+export * from './_colors';
+export * from './_sizes';
 export { defaultThemeOptions } from './theme';
